refactor(AVP): extract GrandStaff notation into protocol-driven MVVM
Extract rendering logic into GrandStaffNotationRenderer (protocol-backed),
move presentation state into GrandStaffNotationPresentationViewModel.
Remove monolithic GrandStaffNotationView+StemsAndBeams.swift.
