+++ LonelyPianistAVP/ViewModels/PracticeNotation/GrandStaffNotationPresentation.swift
@@ -1,0 +1,11 @@
+import CoreGraphics
+
+struct GrandStaffNotationPresentation {
+    let notationLayout: GrandStaffNotationLayout
+    let viewportLayout: GrandStaffNotationViewportLayoutService.Layout
+    let chordsByID: [String: GrandStaffNotationChord]
+    let itemsByChordID: [String: [GrandStaffNotationItem]]
+    let beamedChordIDs: Set<String>
+    let ledgerStepsByItemID: [String: [Int]]
+    let defaultScrollAnchorY: CGFloat
+}
