+++ LonelyPianistAVP/ViewModels/PracticeNotation/GrandStaffNotationPresentationViewModelProtocol.swift
@@ -1,0 +1,13 @@
+import CoreGraphics
+
+protocol GrandStaffNotationPresentationViewModelProtocol {
+    func makePresentation(
+        size: CGSize,
+        lineSpacing: CGFloat,
+        guides: [PianoHighlightGuide],
+        currentGuide: PianoHighlightGuide?,
+        measureSpans: [MusicXMLMeasureSpan],
+        context: GrandStaffNotationContext?,
+        scrollTick: Double?
+    ) -> GrandStaffNotationPresentation
+}
