+++ LonelyPianistAVP/Views/Practice/GrandStaffNotationRendererProtocol.swift
@@ -1,0 +1,9 @@
+import SwiftUI
+
+protocol GrandStaffNotationRendererProtocol {
+    func draw(
+        presentation: GrandStaffNotationPresentation,
+        in context: GraphicsContext,
+        displayScale: CGFloat
+    )
+}
